Save flash by dropping decimals

+++ Software/src/devboard/webserver/webserver.cpp
@@ -1756,9 +1756,10 @@ void onOTAProgress(size_t current, size_t final) {
   // Log every 1 second
   if (ota_progress_timer.elapsed()) {
     if (final > 0) {
-      constexpr float BYTES_PER_KB = 1024.0f;
-      float percent = (float)current * 100.0f / (float) final;
-      logging.printf("OTA progress: %.1f%% (%.1f / %.1f KB)\n", percent, current / BYTES_PER_KB, final / BYTES_PER_KB);
+      constexpr int BYTES_PER_KB = 1024;
+      int percent = current * 100 / final;
+      logging.printf("OTA progress: %u%% (%u / %u KB)\n", percent, (unsigned)(current / BYTES_PER_KB),
+                     (unsigned)(final / BYTES_PER_KB));
     }
     // Reset the "watchdog"
     ota_timeout_timer.reset();
